Network reprezentation

+++ inc/synchronization_request.h
@@ -7,11 +7,11 @@ class synchronization_request {
 	public:
 		uint32_t time;
 		uint32_t pid;
-		uint8_t tag;
+		uint16_t tag;
 
 		synchronization_request();
 		synchronization_request(char*);
-		synchronization_request(uint32_t time, uint32_t pid, uint32_t tag);
+		synchronization_request(uint32_t time, uint32_t pid, uint16_t tag);
 
 		bool operator>(const synchronization_request& request) const;
